# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Canvas App Backend API
  version: 1.1.0
  description: |
    OpenAPI contract for Canvas App backend.
    This version includes all currently implemented routes: users, folders,
    canvas, storage, friends, sharing, and layouts.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://mycanvas-app-backend.vercel.app
    description: Production (Vercel)
...
paths:
...
  /api/folders:
    get:
      tags: [Folders]
      summary: List user folders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Folders fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderListResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Folders]
      summary: Create folder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderSingleResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/canvas:
    get:
      tags: [Canvas]
      summary: List root-level canvases
      description: Returns current user's canvases where folderId is null.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Canvas list fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasListResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Canvas]
      summary: Create canvas
      description: Creates a new canvas in root or inside a folder.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCanvasRequest"
      responses:
        "201":
          description: Canvas created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanvasSingleResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User or folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/friends:
    get:
      tags: [Friends]
      summary: List friends
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Friends fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericDataResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/layouts/canvas:
    get:
      tags: [Layouts]
      summary: Get all canvas layouts for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Canvas layouts fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenericDataResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Request failed
        error:
          type: string
          nullable: true
          example: Internal Server Error
...
    GenericDataResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          nullable: true
          example: Data fetched successfully
        data:
          type: object
          additionalProperties: true
          nullable: true
...
    Folder:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        positionX:
          type: number
          nullable: true
        positionY:
          type: number
          nullable: true
    FolderListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Folders fetched successfully
        folders:
          type: array
          items:
            $ref: "#/components/schemas/Folder"
    FolderSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Folder fetched successfully
        folder:
          $ref: "#/components/schemas/Folder"
    Canvas:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        size:
          type: integer
          example: 0
        url:
          type: string
          example: /canvas/abc123
        folderId:
          type: string
          nullable: true
        ownerId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        positionX:
          type: number
          nullable: true
          example: 120
        positionY:
          type: number
          nullable: true
          example: 80
...
    XYPosition:
      type: object
      required: [x, y]
      properties:
        x:
          type: number
          example: 120
        y:
          type: number
          example: 80
...
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          example: Product Docs
        position:
          $ref: "#/components/schemas/XYPosition"
...
    CreateCanvasRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          example: Feature Board
        folderId:
          type: string
          nullable: true
          example: folder_123
...
    CanvasListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Canvas fetched successfully
        canvas:
          type: array
          items:
            $ref: "#/components/schemas/Canvas"
    CanvasSingleResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Canvas details fetched successfully
        canvas:
          $ref: "#/components/schemas/Canvas"
...